OneDrive/Pulpit/yt-dashboard — kopia/requirements.txt
streamlit
pandas
requests
isodate
python-dotenv